==> projects/main/code/07b_plot_zonal_int.R <==
library(raster)
library(data.table)
library(ggplot2)
library(viridis)
library(dplyr)
#library(reshape)
#library(terra)
library(ncdf4)
library(sf)
library(hms)
library(forcats)
library(parallel)

#source('./source/libs.R')
source('./source/themes.R')
source('./source/palettes.R')
source('./source/graphics.R')


#########################################


data_list <-  readRDS("./projects/main/data/hourly_int_thres_0.1_0.5_all_datasets_LST_glob_2001_20.rds")


zonmean_data_list <- lapply(data_list, function(df) df[, .('0.1' = mean(prec_int, na.rm = TRUE), 
                                                           '0.2' = mean(prec_int_0.2, na.rm = TRUE),
                                                           '0.5' = mean(prec_int_0.5, na.rm = TRUE)), by = .(lat, name)])

zonmean_data <- rbindlist(zonmean_data_list)
to_plot <- melt(zonmean_data,  c("lat", "name"), variable.name = "threshold")


levels(to_plot$threshold) <- c("0.1 (mm/hr)", "0.2 (mm/hr)", "0.5 (mm/hr)")
levels(to_plot$name) <- c("IMERG", "GSMaP", "CMORPH", "PERSIANN", "ERA5")

to_plot[value > 3]
summary(to_plot)
summary(to_plot[lat >= -58.875 & lat <= 58.875])

ggplot(to_plot[lat >= -58.875 & lat <= 58.875], aes(lat, value, col = name), size = 0.5) + 
  geom_line() + 
  scale_color_manual(values = line_colors) + 
  labs(x = "Latitude", y = "Intensity (mm/hr)", col = " ") + 
  theme_generic + 
  facet_wrap(~threshold) + 
  theme(strip.background = element_rect(fill = "white"),
        strip.text = element_text(colour = 'Black'))

ggsave("./projects/main/results/07b_zonal_int_lineplot_thres_0.1_0.5_glob.png",
       width = 9.0, height = 5.2, units = "in", dpi = 600)

##flpped x-y axix----

ggplot(to_plot[threshold == "0.1 (mm/hr)" & lat >= -58.875 & lat <= 58.875], aes(lat, value, col = name), size = 0.5) + 
  geom_line() + 
  scale_color_manual(values = line_colors) + 
  labs(x = "Latitude", y = "Intensity (mm/hr)", col = " ") + 
  theme_generic + 
  #facet_wrap(~threshold) + 
  coord_flip() + 
  scale_x_continuous(breaks = seq(-60, 60, by=10), expand = c(0, 0)) + 
  theme(strip.background = element_rect(fill = "white"),
        strip.text = element_text(colour = 'Black'), legend.position = "right", legend.direction = "vertical") + 
  theme(legend.position = c(0.8, 0.2))


ggsave("./projects/main/results/07b_zonal_int_lineplot_flip.png",
       width = 4.5, height = 8.8, units = "in", dpi = 600)


### for land ocean and globe-------

zonmean_data_list_landocn <- lapply(data_list, function(df) df[, .(#'no_threshold' = mean(prec_int, na.rm = TRUE), 
                                                                   '0.1' = mean(prec_int, na.rm = TRUE), 
                                                                   '0.2' = mean(prec_int_0.2, na.rm = TRUE),
                                                                   '0.5' = mean(prec_int_0.5, na.rm = TRUE)), by = .(lat, name, location)])

zonmean_data_landocn <- rbindlist(zonmean_data_list_landocn)
to_plot_landocn <- melt(zonmean_data_landocn,  c("lat", "name", "location"), variable.name = "threshold")
to_plot_landocn <- to_plot_landocn[, .(lat, name, threshold, value, location)]
levels(to_plot_landocn$threshold) <- c("0.1 (mm/hr)", "0.2 (mm/hr)", "0.5 (mm/hr)")
levels(to_plot_landocn$name) <- c("IMERG", "GSMaP", "CMORPH", "PERSIANN", "ERA5")
to_plot_glob <- to_plot[, `:=`(location = factor("global"))]

to_plot_landocnglob <- rbind(to_plot_glob, to_plot_landocn)
levels(to_plot_landocnglob$location) <- c("Global", "Land", "Ocean")


ggplot(to_plot_landocnglob[lat >= -58.875 & lat <= 58.875], aes(lat, value, col = name), size = 0.5) + 
  geom_line() + 
  scale_color_manual(values = line_colors) + 
  labs(x = "Latitude", y = "Intensity (mm/hr)", col = " ") + 
  theme_generic + 
  facet_grid(location~threshold) + 
  theme(strip.background = element_rect(fill = "white"),
        strip.text = element_text(colour = 'Black'))

ggsave("./projects/main/results/07b_zonal_int_lineplot_thres_0.1_0.5_landocnglob.png",
       width = 9.0, height = 4.8, units = "in", dpi = 600)



#######################################################################

## have to check theinput data first
# for seasonal -------------------------------------------------

data_list <-  readRDS("./projects/main/data/hourly_int_thres_0.1_0.5_all_datasets_LST_glob_2001_20_seasonal.rds")


zonmean_data_list <- lapply(data_list, function(df) df[, .(prec_int = round(mean(prec_int, na.rm = TRUE), 2)), by = .(lat, name, threshold, season)])

zonmean_data <- rbindlist(zonmean_data_list)
to_plot <- zonmean_data


levels(to_plot$threshold) <- c("0.1 (mm/hr)", "0.2 (mm/hr)", "0.5 (mm/hr)")
to_plot$name <- factor(to_plot$name, levels = c("imerg", "gsmap", "cmorph", "persiann", "era5"))
levels(to_plot$name) <- c("IMERG", "GSMaP", "CMORPH", "PERSIANN", "ERA5")
levels(to_plot$season) <- c("JJA", "DJF")

to_plot[prec_int > 30]

ggplot(to_plot[prec_int < 4], aes(lat, prec_int, col = name), size = 0.5) + 
  geom_line() + 
  scale_color_manual(values = line_colors) + 
  labs(x = "Latitude", y = "Intensity (mm/hr)", col = " ") + 
  theme_generic + 
  facet_grid(season~threshold) + 
  theme(strip.background = element_rect(fill = "white"),
        strip.text = element_text(colour = 'Black'))

ggsave("./projects/main/results/07b_zonal_int_lineplot_thres_0.1_0.5_glob_seasonal.png",
       width = 9.0, height = 5.2, units = "in", dpi = 600)


### for land ocean and globe------------------------------------------------------

zonmean_data_list_landocn <- lapply(data_list, function(df) df[, .(prec_int = round(mean(prec_int, na.rm = TRUE), 2)), by = .(lat, name, threshold, season, location)])

zonmean_data_landocn <- rbindlist(zonmean_data_list_landocn)
#to_plot_landocn <- melt(zonmean_data_landocn,  c("lat", "name", "location"), variable.name = "threshold")
to_plot_landocn <- zonmean_data_landocn[, .(lat, name, threshold, season, prec_int, location)]
levels(to_plot_landocn$threshold) <- c("0.1 (mm/hr)", "0.2 (mm/hr)", "0.5 (mm/hr)")
to_plot_landocn$name <- factor(to_plot_landocn$name, levels = c("imerg", "gsmap", "cmorph", "persiann", "era5"))
levels(to_plot_landocn$name) <- c("IMERG", "GSMaP", "CMORPH", "PERSIANN", "ERA5")
levels(to_plot_landocn$season) <- c("JJA", "DJF")
to_plot_glob <- to_plot[, `:=`(location = factor("global"))]

to_plot_landocnglob <- rbind(to_plot_glob, to_plot_landocn)
levels(to_plot_landocnglob$location) <- c("Global", "Land", "Ocean")
levels(to_plot_landocnglob$threshold) <- c("0.1", "0.2", "0.5")
levels(to_plot_landocnglob$season)

ggplot(to_plot_landocnglob[prec_int < 3], aes(lat, prec_int, col = name), size = 0.5) + 
  geom_line() + 
  scale_color_manual(values = line_colors) + 
  labs(x = "Latitude", y = "Intensity (mm/hr)", col = " ") + 
  theme_small + 
  #facet_grid(threshold~season~location) + 
  facet_grid(location~season~threshold) +
  theme(strip.background = element_rect(fill = "white"),
        strip.text = element_text(colour = 'Black'))

ggsave("./projects/main/results/07b_zonal_int_lineplot_thres_0.1_0.5_landocnglob_seasonal.png",
       width = 9.0, height = 4.8, units = "in", dpi = 600)
